+++ content/themes/default/page.php
@@ -21,7 +21,7 @@ get_header(['page' => $page]);
             <h1 class="lp-page-title"><?= esc_html($page->title) ?></h1>
             <?php if (has_post_thumbnail($page)): ?>
                 <div class="lp-post__thumbnail lp-gallery">
-                    <?php the_post_thumbnail_lightbox($page, size: 'large', largeSize: 'large'); ?>
+                    <?php the_post_thumbnail_lightbox($page, size: 'large'); ?>
                 </div>
             <?php endif; ?>
             <div class="lp-post__content"><?= render_content($page->content, $page->contentFormat) ?></div>